Initial commit for LevelDesign
Kojima Island has been imported as a base for usage on implementing the current track design, however due to a light corruption progress has been restarted. As well as this, with the introduction of car controls the proportion and scale of the mountain and initial slab of track have been constantly increased in order to match.