Current Elimination Progress
Elimination has been started on, methods to have an object disappear based on their visibility being looked into. Currently though OnBecomeInvisible looks to be the biggest potential solution, it seemingly works differently in the Unity editor compared to a compiled game and thus a different solution may need to be found. This commit's purpose is to push all worked on files in order to merge with the current camera work, in hopes of it clashing better and resulting in a successful implementation.