Elimination Method Implemented
As of now, a firm method for elimination has been found and implemented. Having used the recently created camera system the boundary based method worked and a list of player objects found through tags was used in order to sort through them, however on further investigation the RaceManager was found and used as it would link with the overall project better and possibly be easier to work with. 

From here the system was found to erase the first player that exited the camera with most cases being the frontrunner. In order to combat this and ensure players were only eliminated fairly, a timer was to be implemented however using an ienumerator seemed to loop with itself and its calling while creating no appropriate result, likely due to a logic error. 

Eventually, the idea of an ienumerator was shed and instead a deltaTime based timer was created. This uses a time limit float which would be reached with timers for each vehicle, then dealing with the consequences correctly afterwards. At first however, this did not seem to be the outcome and this may have been due to each if statement checking the boundaries each time despite being inside of a valid boundary check. In all, this and the issue pertaining to the ienumerator could be assumed to be due to loop based logic, in which the code acted as if each vehicle checked were its own separate frame. As such, the checks in this circumstance would prove to be too much and as a result heavily slow the code itself. Removing these checks have not only notably streamlined the code, but also resulted in the intended outcome. From here, a cleaner method of destruction will be implemented in order to comply with the rest of the codebase and ensure that objects are not erased despite needing to be used for other purposes.