Added timer reset
As a small realisation tested out to be true, having the timer without a reset has it act more like HP than a timer in which players can still return to the race. As a result, a reset has been added so that the amount of time lost from being out of view doesn't transfer into the next time it happens.