Race Winner and Name Tag Progress
Detecting a winner has been implemented, an eliminated variable becoming true when a vehicle deactivates and another total being brought up. If the number of eliminated vehicles is equal to the total minus 1, the remaining car is given the winner.

Currently for the Name Tag script, GameObjects are created in the canvas in order to hold the TextMeshPro feature, adding the vehicle's name to itself. For making itself shown, efforts still need to be placed into correctly implementing this as at the moment, the text objects are placed further away despite being converted to the screenpoint.