Nametag Progress
With a set offset and font size, the current target destination for the name tags uses the direct positions of each vehicle instead of translating them to screen point co-ordinates. This has brought far more success than before, and the offset has been configured so that the tags are placed correctly. As for current flukes, once the Player 1 vehicle leaves the others it adopts the Player 2 tag. This seems to be due to the fact that the Player 2 vehicle is the first processed car in the carSystem, in turn creating its nametag first and by extension applying it to Player 1  due to its order. This may require discussion with those involved with processing the vehicles, as it would seem this is an externally arisen issue that the Name Tag script is not at fault for.