Nametag and Elimination Integration for Level Design Scene
Initially the nametags have been arranged for the LevelDesign scene, including appropriate offsets, sizes and rotations. As to allow for easier viewing, outlines were worked on for the tags and after seeing a lack of any after specifying a colour and width, research found that a specific shader may have been required. Many methods were used including Shader.Find, Resources.Load, easier access to the appropriate shader and even creation of a clone font to set the shader easier however these unfortunately did not work. Later on, it was realised that the TMP's own material and shader was being looked into for appropriate alteration and not the material and shader of the font, so this was corrected and finally with Shader.Find the intended outcome has been reached. As well as this, the Elimination Script works as mostly intended, with just a small amount of the visible screen boundaries counting as outside. This will be looked into and solved if it proves to become more of an issue.