Nametag and Player Identification Progress
Name tags have now been stylised partially, with a custom texture being created for them. The texture uses expressive vibrant colouring to match with the game's aesthetic while helping the text be more pronounced in gameplay, also using the TextMeshPro Distance Field shader over the Mobile edition so that outlines can be visualised. 

However, discussions have noted the viability of name tags for player identification in a local setting not to mention the screen already requiring many UI aspects, possibly concluding that they may be unnecessary for the game's current state and  not quick enough an identifier. As such, a more fitting local implementation of player identification has been implemented in the form of different player vehicle colours, being notably easier and simpler for the mind to read and consequently being a more efficient method. 

As such, the roofs of the vehicles are allocated different colours for each player in script through the alteration of the vehicle model's roof material so that each number of players is accommodated for.

Due to the model's structure and the parent gameobject, the model itself is a sibling of the PositionUpdate focused object that the script accessed through a GetChild statement for the parent. Once the material was also declared, the colour is then changed through the player number with a switch case filtering this. Though implementations such as material.color and material.SetTexture did not provide a result, changing the shader/material's "_BaseColor" through material.setColor managed to show the intended outcome. With the successful implementation, this method of identification has been segmented away from its original conceptualisation in NameTagScript and thus has been given its own script specifically.