Elimination Total Fix
Concerns were raised over the winner state not being reached, and the root cause was quickly found to be the eliminatedTotal variable. A check was added for already eliminated vehicles to add to the variable in case of a new loop, with it resetting to 0 each new loop. However, with a circumstance such as Player 2 being eliminated, the total would reset to 0 and thus once Player 1 would be processed the winner if statement would not be satisfied as the eliminatedTotal would be too low. Thus, eliminatedTotal has been made a global variable and the refresh satisfaction check has been removed, which should fix the issue.