UI Bars Progress
Basic UI bars have been added. A prefab canvas has been created holding four bars, each with a custom made border and inner bar (which has been given a slider property as to fit appropriately). Currently they also hold text objects, but these are to be implemented further once a solid ranking can be accessed. Through a new Bar script, these objects as well as a gradient are declared and allocated. With the slider manipulation, the poweramount from the Abilities script in each vehicle is used and alongside this, also affects the slider's colour due to the aforementioned use of a Gradient. However, this only currently shows one colour at a time going from red to orange to yellow, and a bar with the gradient flowing through it is more desired. This is being heavily looked into, and should be implemented soon alongside proper ranking portrayal.