UI Bar Text Progress
A work in progress ranking system is now present in order to assist with the UI bar text. This is currently done through looping through each vehicle and comparing collider hit and lap count amounts to the other players. Since the colliders count as checkpoints on the track, hitting more would on average result in the player with a higher count being placed higher above the others and thus this increments an "aheadOf" variable which keeps track of the number of vehicles the current is ahead of. Once this number is given to a variable, it is then found which position it corresponds to in relation to the number of present vehicles. Initially, a switch case was intended for this logic but was removed due to the syntax not being able to use non static values, and an intended enumerator was also removed due to being made redundant through the already implemented comparison method. Currently, an issue has been found causing only the top left text to read 1st even if Player 1 is not leading, as well as the consumption of other players' bars by unintended players in a number of circumstances. This is likely due to the RaceManager swapping the positions in the vehicle array, but more investigation into the matter may bring about a more firm cause.