Cleanup of unneeded code and bar colour method
As a whole, a uniform colour bar has been decided by the group as the most fitting for the project, so work on implementing a visual gradient esque bar has been dropped. To add, Color.Lerp results in the same outcome as the lengthy setup and use of a gradient, so this has also been replaced as to be more proficient and clean up code.