Preparation for Fleshed Out Elimination
Preparation is currently underway to expand on and flesh out the elimination system. Each round and eventually game winner will be passed through the GameManager Instance as it holds all relevant instance dependant information, however progress requires the tying in of player IDs and the current car in order to be consistently correct. A test value is included to show the correct base implementation but flaw with using an inaccurate value.