Elimination System Progress
Minor changes: VehicleColours has been unchecked to halt errors, attempts to have it function alongside the new TukTuk models are not entirely recommended as the model and its structure structure seems to change very often and as of current different coloured models have been created.

As of current, the implementation has been done through attaching Core Car Module to the "Auto Rickshaw" component of each vehicle and altering appropriately. The attachment helps the system acquire the player ID, which is then correlated to the m_Players array in GameManager and the round winner is submitted through the very same class. As well as this, the game winner is also submitted once the total number of rounds won is found to be a multiple of 3. Initially, tests done without breakpoints would add two round wins per win, seemingly due to the code eliminating the eliminated player(s) again and thus fulfilling the gate quota of telling the system the remaining player had won. This was fixed via a "processed winner" variable that closes off the elimination process if a winner has already been processed. As the script is currently reset each new game due to the switching of scenes, this allows for an automatic toggle which helps restrain proceedings.