Round Victory Juice Progress
Currently, work is being done towards an animation and intermission for round victories. The initial idea involved the flowing in of a banner alongside the winning vehicle driving in with an angle, followed by their round wins represented by stars. 

Attempts were made to replicate this, firstly with visualising the models in the Canvas space. These proved to be unfortunately fruitless however, and it was eventually found that such a thing would not work in HDPR mostly due to its emphasis on the usage of one camera per scene. 

Thus, 2D sprites have been created for the implementation and will sweep across the banner unveiling the round wins for that player. The UIBarCanvas prefab has been updated to include the elements to be used, and animations have been created for the TukTuk and banner. Finally, an animation script has been created to implement this logic, and various others have been modified to better suit the circumstances (ResetTrack will be used to call the animation functions). 

This commit has been pushed in order to give way for merging of the Event system, so that easier implementation can be made for both the round animation and such factors as removing player input for the animation and placing players at specific checkpoints at the beginning of each round.