Continued Animation Progress
A new Elimination scene has been created in order to move progress onto the new style of scene. As well as this, temporary star sprites have been created and the skeleton of the animation script has been fleshed out. Currently, a default win number is assigned and the IEnumerator toggles the Animator booleans to create the animation.